Ввел общий вид ответа, в идеале надо будет вынести в стартер Отрефакторил тесты Продолжил делать заявки

+++ repositories/src/main/java/by/mmb/repo/CityRepository.java
@@ -12,5 +12,5 @@ public interface CityRepository {
 
     boolean addNewCity(City city);
     City getCityById(long idCity);
-    boolean isPresent(List<Integer> ids) throws AppsException;
+    boolean isPresent(List<Long> ids) throws AppsException;
 }
